Refactor project name to match mod name
