+++ gradle.properties
@@ -9,8 +9,8 @@ yarn_mappings=1.21.5+build.1
 loader_version=0.16.10
 
 # Mod Properties
-mod_version=1.0.1
-maven_group=com.example
+mod_version=1.0.2
+maven_group=ca.servermetrics
 archives_base_name=servermetrics
 
 # Dependencies
