+++ src/client/java/ca/servermetrics/FabricMainClient.java
@@ -1,8 +1,8 @@
-package com.example;
+package ca.servermetrics;
 
 import net.fabricmc.api.ClientModInitializer;
 
-public class ExampleModClient implements ClientModInitializer {
+public class FabricMainClient implements ClientModInitializer {
 	@Override
 	public void onInitializeClient() {
 		// This entrypoint is suitable for setting up client-specific logic, such as rendering.
